+++ stories/components/alert-drawer/alert-drawer.stories.ts
@@ -27,8 +27,8 @@ export default meta;
 
 let alertId = 1;
 
-const addAlert = (wcsAlertIntent: WcsAlertIntent, showProgressBar: boolean, timeout: number) => {
-    const alertDrawer = document.getElementById("wcs-alert-drawer");
+const addAlert = (wcsAlertIntent: WcsAlertIntent, showProgressBar: boolean, timeout: number, alertDrawerId: string) => {
+    const alertDrawer = document.getElementById(alertDrawerId);
     if (!alertDrawer) {
         return;
     }
@@ -47,30 +47,44 @@ const addAlert = (wcsAlertIntent: WcsAlertIntent, showProgressBar: boolean, time
     }
 };
 
+const clearAlerts = (alertDrawerId: string) => {
+    const alertDrawer = document.getElementById(alertDrawerId);
+    if (!alertDrawer) {
+        return;
+    }
+
+    alertDrawer?.clear();
+}
+
 type AlertDrawerArgs = {
     position: WcsAlertDrawerPosition,
     showProgressBar: boolean,
     timeout: number,
+    alertDrawerId: string
 }
 
-const Template: StoryFn<AlertDrawerArgs> = (args) => html`
-    <p>💡Use attributes in the panel below to play with <b>wcs-alert-drawer</b> and <b>wcs-alert</b></p>
+const Template: StoryFn<AlertDrawerArgs> = (args, context) => {
+    const uniqueDrawerId = `${args.alertDrawerId}-${context.id}`;
 
-    <wcs-button @click="${() => addAlert('success', args.showProgressBar, args.timeout)}" class="wcs-success">Toggle
-        success alert
-    </wcs-button>
-    <wcs-button @click="${() => addAlert('information', args.showProgressBar, args.timeout)}" class="wcs-info">Toggle
-        information alert
-    </wcs-button>
-    <wcs-button @click="${() => addAlert('warning', args.showProgressBar, args.timeout)}" class="wcs-warning">Toggle
-        warning alert
-    </wcs-button>
-    <wcs-button @click="${() => addAlert('error', args.showProgressBar, args.timeout)}" class="wcs-critical">Toggle
-        error alert
-    </wcs-button>
+    return html`
+        <p>💡Use attributes in the panel below to play with <b>wcs-alert-drawer</b> and <b>wcs-alert</b></p>
 
-    <wcs-alert-drawer position=${args.position} id="wcs-alert-drawer"></wcs-alert-drawer>
-`
+        <wcs-button @click="${() => addAlert('success', args.showProgressBar, args.timeout, uniqueDrawerId)}" class="wcs-success">Toggle
+            success alert
+        </wcs-button>
+        <wcs-button @click="${() => addAlert('information', args.showProgressBar, args.timeout, uniqueDrawerId)}" class="wcs-info">Toggle
+            information alert
+        </wcs-button>
+        <wcs-button @click="${() => addAlert('warning', args.showProgressBar, args.timeout, uniqueDrawerId)}" class="wcs-warning">Toggle
+            warning alert
+        </wcs-button>
+        <wcs-button @click="${() => addAlert('error', args.showProgressBar, args.timeout, uniqueDrawerId)}" class="wcs-critical">Toggle
+            error alert
+        </wcs-button>
+
+        <wcs-alert-drawer position=${args.position} id=${uniqueDrawerId}></wcs-alert-drawer>
+    `;
+}
 
 export const Default: StoryObj<AlertDrawerArgs> = {
     render: (args: AlertDrawerArgs, context) => Template(args, context),
@@ -78,5 +92,33 @@ export const Default: StoryObj<AlertDrawerArgs> = {
         position: 'bottom-right',
         timeout: 5000,
         showProgressBar: true,
+        alertDrawerId: 'wcs-alert-drawer'
+    }
+};
+
+/**
+ * You can use `clear()` method exposed on `wcs-alert-drawer` to clear all alerts of the `wcs-alert-drawer`
+ */
+export const ClearMethod: StoryObj<AlertDrawerArgs> = {
+    render: (args, context) => {
+        const uniqueDrawerId = `${args.alertDrawerId}-${context.id}`;
+
+        return html`
+            <p>💡Use attributes in the panel below to play with <b>wcs-alert-drawer</b> and <b>wcs-alert</b></p>
+            <wcs-button @click="${() => addAlert('success', args.showProgressBar, args.timeout, uniqueDrawerId)}" class="wcs-success">
+                Toggle success alert
+            </wcs-button>
+            <wcs-button class="wcs-critical" @click="${() => clearAlerts(uniqueDrawerId)}" class="wcs-success">
+                Clear alerts
+            </wcs-button>
+
+            <wcs-alert-drawer position=${args.position} id=${uniqueDrawerId}></wcs-alert-drawer>
+        `;
+    },
+    args: {
+        position: 'bottom-right',
+        timeout: 5000,
+        showProgressBar: true,
+        alertDrawerId: 'wcs-alert-drawer-clear'
     }
 };