+++ CHANGELOG.md
@@ -28,6 +28,7 @@ for now removed features.
 ### Fixed
 
 - **Angular**: set `compilationMode` to `partial` for `wcs-angular` and `wcs-formly` libraries to fix `AsyncPipe` import issues in consuming applications
+- **alert-service**: fix `clear` method which did not clear alerts
 
 ### Security